+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- upgrade to Cloudfeaster 0.9.5
+- upgrade to [Cloudfeaster 0.9.7](https://github.com/simonsdave/cloudfeaster/releases/tag/v0.9.7)
 - changed miniclip spider to be a bit smarter and figure out how many
   games where in the top N list rather than assuming it was always 10 games
 - pep8 -> pycodestyle