+++ CHANGELOG.md
@@ -19,7 +19,7 @@ or inside a docker container
 
 ### Removed
 
-- Nothing
+- removed ```Xvfb``` from ```cfg4dev```
 
 ## [1.1.0] - [2017-11-28]